+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.18.0 (2022-10-21)
 
+### Added
+- `setState` and `setStateQuiet` with `transform` lambda argument.
+
 ### Changed
 - Deprecate `onTransition`, replace with `onEnter`.
 - Make it impossible to nest triggers.